Rename internal flask package
