+++ localtileserver/client.py
@@ -4,8 +4,8 @@ from typing import List, Union
 
 import requests
 
-from localtileserver.application.palettes import palette_valid_or_raise
-from localtileserver.application.utilities import get_clean_filename
+from localtileserver.tileserver.palettes import palette_valid_or_raise
+from localtileserver.tileserver.utilities import get_clean_filename
 from localtileserver.utilities import add_query_parameters, save_file_from_request
 
 DEMO_REMOTE_TILE_SERVER = "https://localtileserver-demo.herokuapp.com/"
