+++ localtileserver/tileserver/palettes.py
