+++ localtileserver/tileserver/static/js/cesium.js
