+++ localtileserver/tileserver/static/js/geojsControls.js
