+++ localtileserver/tileserver/static/styles/cesium.css
