+++ localtileserver/tileserver/templates/tileserver/_include/cesium.html
