+++ localtileserver/tileserver/templates/tileserver/_include/geojs.html
