+++ localtileserver/tileserver/templates/tileserver/_include/geojsControls.html
