+++ localtileserver/tileserver/views.py
@@ -3,8 +3,8 @@ import logging
 from flask import current_app, render_template, request
 from flask.views import View
 
-from localtileserver.application import utilities
-from localtileserver.application.blueprint import tileserver
+from localtileserver.tileserver import utilities
+from localtileserver.tileserver.blueprint import tileserver
 from localtileserver.examples import get_data_path
 
 logger = logging.getLogger(__name__)
